Fix counter scope

+++ src/desugarer.rs
@@ -5,7 +5,7 @@ use crate::parser_helpers::{
 
 pub struct Desugarer {
     pub json: String,
-    pub last_id: u32,
+    pub last_id: usize,
 }
 
 pub struct ParagraphIndentOptions {
@@ -24,7 +24,7 @@ pub struct IgnoreOptions {
 }
 
 impl Desugarer {
-    pub fn new(json: &str, last_id: u32) -> Desugarer {
+    pub fn new(json: &str, last_id: usize) -> Desugarer {
         Desugarer {
             json: json.to_string(),
             last_id,
@@ -368,7 +368,7 @@ impl Desugarer {
     }
 
     pub fn is_first_child(
-        element_id: u32,
+        element_id: usize,
         parent: &Option<&mut DataCell>,
         options: &ParagraphIndentOptions,
     ) -> bool {
@@ -418,7 +418,7 @@ impl Desugarer {
         false
     }
 
-    pub fn prev_is_delimited(element_id: u32, parent: &Option<&mut DataCell>) -> bool {
+    pub fn prev_is_delimited(element_id: usize, parent: &Option<&mut DataCell>) -> bool {
         if let Some(parent) = parent {
             if let CellType::Element(parent) = &parent.cell_type {
                 let mut prev_el: Option<&DataCell> = None;
@@ -437,7 +437,7 @@ impl Desugarer {
     }
 
     pub fn tags_before_non_indents(
-        element_id: u32,
+        element_id: usize,
         parent: &Option<&mut DataCell>,
         option: &ParagraphIndentOptions,
     ) -> bool {
